refactor(treepad): replace Service facade with package-level functions + Deps
Removes the stateless Service struct and splits internal/treepad into one
file per operation (generate, new, cd, remove, prune, status). Shared
helpers extracted into dedicated files:

- deps.go: Deps struct + DefaultDeps(out, in) — single composition root
  that fixes the 6× duplicated NewService wiring across commands and
  eliminates workspace.go's silent nil Opener
- repo_context.go: loadRepoContext — collapses the repeated
  list→main→slug→outputDir prologue from 5 call sites into one
- remove_helper.go: removeWorktreeAndArtifact(force bool) — collapses
  removeWorktree / forceRemoveWorktree near-duplicates

Tests split per op into generate/new/cd/remove/prune/status_test.go;
shared fakes consolidated into testfakes_test.go.

Closes #43
